pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-fast-treenode"
version = "3.0.0"
description = "Treenode Framework for supporting tree (hierarchical) data structure in Django projects"
readme = "README.md"
authors = [{ name = "Timur Kady", email = "timurkady@yandex.com" }]
license = { file = "LICENSE" }
requires-python = ">=3.9"
dependencies = [
    'Django>=4.0',
    'msgpack>=1.0.0',
    'openpyxl>=3.0.0',
    'pyyaml>=5.1',
]
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Intended Audience :: Developers',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3.13',
    'Programming Language :: Python :: 3.14',
    'Framework :: Django',
    'Framework :: Django :: 4.0',
    'Framework :: Django :: 4.1',
    'Framework :: Django :: 4.2',
    'Framework :: Django :: 5.0',
    'Framework :: Django :: 5.1',
    'Framework :: Django :: 5.2',
    'License :: OSI Approved :: MIT License',
    'Operating System :: OS Independent',
]

[project.urls]
Homepage = "https://github.com/TimurKady/django-fast-treenode"
Documentation = "https://django-fast-treenode.readthedocs.io/"
Source = "https://github.com/TimurKady/django-fast-treenode"
Issues = "https://github.com/TimurKady/django-fast-treenode/issues"

